Fix 5806: Battery usage data not available
PowerUsageSummary activity is reporting "data not available" because levels defined in power_profile.xml
are lower than 10 for POWER_SCREEN_FULL.
The actual power_profile.xml is the original aosp file for be customized for every device.
Jellybean branch has an update versión of this file.

This change add this jellybean power_profile.xml file to ICS.

Change-Id: Ifbfb4bf966e65996291b72d29224f3fcdf890417