Update patch set 1
Patch Set 1:

Well, there is no cm-10.1 branch (yet) in device/asus/tf101 ...

Yes, with that patch alone the tf101 doesn't boot yet.  Not because of the EGL settings, but because the storage setup must be fixed for 4.2 multiuser.  There are a few more patches because of that, look here:

http://review.cyanogenmod.org/#/q/status:open+project:CyanogenMod/android_device_asus_tf101,n,z

Builds are here: http://www.kraxel.org/cm/tf101/

Patch-set: 1