Update patch set 1
Patch Set 1:

I believe you may be able to submit it to the right branch by using git checkout, but ask around first.

I'm just interested in keeping the branches separate so people are still able to build for cm10 if they so wish.

I realise that this might be a pain for the submitters though, when you have half a dozen changes open.

This is just my opinion though, ask around on IRC what the stance is on this.

Patch-set: 1