Update patch set 1
Patch Set 1:

I'll do as suggested and report back. I just started building this again because my son started using it, so I haven't been very commited to this device. Fwiw, on my current builds (without this and the other commit), sound works fine on headphones, but only on tablet when not attached to the keyboard dock.

Separately, I'm more concerned by the sleep of death.

Patch-set: 1