Update patch set 1
Patch Set 1:

@Matthias - I noticed that was merged already. When I say "sleep of death" I don't mean the audio, I mean the whole device shuts off. (although last night's build survived to this morning without SOD, yet)

As for audio... headphones work, and tablet speakers work if not docked. This behavior occurs without any cherry-picks at all, and does not improve or worsen with this, or any other current cherry picks. And I have no boot problems whatsoever. 

Not trying to troll, here. Just seeing results that contradict the descriptions and comments I've seen.

Patch-set: 1