Update patch set 1
Patch Set 1:

Not really.  Well, I've found + fixed some of the issues in the oss audio hal.  But recording still doesn't work, so something must still be missing :(

I still hope to figure what it is, but I think switching to the prebuild audio blobs is sensible for the time being.

commit f1050d2bfee2a6aff60b743a48a000755b94d855 is only one half of it though, we must also re-add the blobs to proprietary-files.txt, for example with with "git revert 25f41db1ab456d88391a00953965e54acdaa0243".

Patch-set: 1